added jest wrapper for e2e test
